TGCN/tgcn_pytorch/requirements.txt
numpy
pandas
torch>=1.9.0
scipy
scikit-learn
matplotlib
pyyaml
